requirements.txt
notify-send==0.0.7
Pillow==5.4.1
schedule==0.6.0
twilio==6.24.0
terminaltables==3.1.0
